feat: 서버통신을 위한 레퍼지토리 패턴 적용

+++ android/domain/src/main/java/com/created/domain/model/Study.kt
@@ -1,7 +1,7 @@
 package com.created.domain.model
 
 data class Study(
-    val studyId: Long,
+    val studyId: Int,
     val studyName: String,
     val progressRate: Int,
     val leftDays: Int,
